feat: 개념증명

+++ crenata/core/user/usecases/delete.py
@@ -1,4 +1,5 @@
 from crenata.core.user.domain.repository import UserRepository
+from crenata.core.user.exceptions import UserNotFound
 
 
 class DeleteUserUseCase:
@@ -6,4 +7,7 @@ class DeleteUserUseCase:
         self.user_repository = user_repository
 
     async def execute(self, user_id: int) -> None:
+        user = await self.user_repository.get_user(user_id)
+        if user is None:
+            raise UserNotFound
         await self.user_repository.delete_user(user_id)
